refactor: tighten io.read API surface and simplify internals
Narrows construction visibility for types that should not be user-built,
removes dead defensive code, and collapses 4 InputStream/Path overload
duplicates to one-liners through a shared withReader helper.

- RecordMapping: final, package-private constructor (still publicly
  reachable as a parameter type via MultiMatchStrategy.resolve)
- ReadResult: package-private constructor; only FixedFormatReader.read
  produces instances
- FixedFormatReaderBuilder: declared final
- FixedFormatLineProcessor.parseRecord: dropped no-op <R> parameter and
  misleading wildcard-capture comment
- LinePattern.positions(): dropped wasted defensive clone across a
  package-private boundary
- FixedFormatReader: extracted withReader(InputStream|Path) helpers;
  added comment explaining seed-then-strip preserves registration order
- RecordMappingIndex: typed buckets field as List<Bucket>; removed
  redundant size() > 1 sort guard
- Tests: moved TestRecordMapping and TestReadResult into io.read so
  they can reach the now-package-private constructors
