+++ fixedformat4j/src/main/java/com/ancientprogramming/fixedformat4j/io/read/LinePattern.java
@@ -112,7 +112,7 @@ public final class LinePattern {
   }
 
   int[] positions() {
-    return positions.clone();
+    return positions;
   }
 
   String literal() {
