+++ fixedformat4j/src/test/java/com/ancientprogramming/fixedformat4j/io/read/TestRecordMapping.java
@@ -1,12 +1,10 @@
-package com.ancientprogramming.fixedformat4j.io;
+package com.ancientprogramming.fixedformat4j.io.read;
 
 import com.ancientprogramming.fixedformat4j.annotation.Field;
 import com.ancientprogramming.fixedformat4j.annotation.Record;
-import com.ancientprogramming.fixedformat4j.io.read.LinePattern;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
-import com.ancientprogramming.fixedformat4j.io.read.RecordMapping;
 
 class TestRecordMapping {
